chore: refactor wavelength
