+++ clovers/src/colors.rs
@@ -2,8 +2,6 @@
 
 #[allow(non_snake_case)]
 pub mod D65;
-pub mod photon;
 pub mod photopic;
 
-pub use photon::*;
 pub use photopic::*;
