+++ clovers/src/hitable.rs
@@ -10,10 +10,10 @@ use crate::objects::{GLTFTriangle, GLTF};
 use crate::{
     aabb::AABB,
     bvhnode::BVHNode,
-    colors::Wavelength,
     materials::MaterialTrait,
     objects::{Boxy, ConstantMedium, MovingSphere, Quad, RotateY, Sphere, Translate, Triangle},
     ray::Ray,
+    wavelength::Wavelength,
     Float, Vec3,
 };
